external-siunitx: move siunitx outside of core
