+++ contrib/siunitx/siunitx.ts
@@ -1,0 +1,6 @@
+import katex from "katex";
+import {registerSiunitx} from "./register";
+
+registerSiunitx();
+
+export default katex;
