+++ webpack.common.js
@@ -54,6 +54,10 @@ const targets = [
         name: 'contrib/render-a11y-string',
         entry: './contrib/render-a11y-string/render-a11y-string.ts',
     },
+    {
+        name: 'contrib/siunitx',
+        entry: './contrib/siunitx/siunitx.ts',
+    },
 ];
 
 /**
